# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Ghana Super App (Sankofa) API
  version: "1.0.0"
  description: >
    Maintainability follow-up: this is the first machine-readable description
    of the API surface — previously it lived only in the zod schemas and
    route files themselves. Coverage is intentionally uneven right now:

      - Auth, Rides and Payments (the original core loop, and the modules
        with the deepest test coverage) are fully specified: request/response
        bodies, status codes, auth requirements.
      - Every other module (Admin, and the eight service tiles — Eat, Fix,
        Utilities, Liquidity, Fleet, Property, Travel, plus Drivers,
        Vehicles, Locations, Notifications, Pricing, Ratings, Support, Users)
        is listed with its real path, method, tag and auth requirement, but
        without full request/response schemas yet — filling those in is real,
        traceable work (see the technical-readiness checklist's
        "Maintainability" section), not a placeholder to hide.

    All endpoints are under the `/api/v1` base path shown in `servers`.
    Every response follows one envelope: `{ "success": true, "data": ... }`
    on success, `{ "success": false, "error": { "code", "message" } }` on
    failure — this is enforced by errorHandler.ts, not repeated on every
    operation below.
servers:
  - url: /api/v1
    description: Base path for every route below (see main.ts)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Short-lived access token from POST /auth/verify-otp or
        POST /auth/refresh. Required on every route below except
        /auth/request-otp, /auth/verify-otp, /auth/refresh and
        /payments/webhook (which authenticates by HMAC signature instead —
        see the payments.webhook path and ADR 0002).
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string }
  responses:
...
    RateLimited:
      description: Too many requests — see middleware/rateLimiter.ts for the limiter this route uses
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
security:
  - bearerAuth: []
paths:
...
  /payments/{id}:
    get:
      tags: [Payments]
      summary: Get one payment by id (must belong to the caller)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Payment returned }
        "404": { description: Not found or not owned by caller }
  /payments/webhook:
    post:
      tags: [Payments]
      summary: Payment gateway callback — NOT customer-authenticated
      description: >
        Authenticated by an HMAC-SHA256 signature over the raw request body
        (header `X-Signature`, secret PAYMENT_WEBHOOK_SECRET) — see ADR 0002
        and webhookSignature.ts. This replaced an earlier customer-callable
        `POST /:id/confirm` that let a customer mark their own payment PAID
        with no money having moved.
      security: []
      parameters:
        - name: X-Signature
          in: header
          required: true
          schema: { type: string }
      responses:
        "200": { description: Payment state updated }
        "401": { description: Missing or invalid signature }
        "429": { $ref: "#/components/responses/RateLimited" }

  # ------------------------------------------------------------------
  # Everything below: real paths, not yet fully schema'd — see info.description
  # ------------------------------------------------------------------
...
  /support/tickets/{id}:
    get:
      tags: [Support]
      summary: Get one ticket
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /support/tickets/{id}/messages:
    post:
      tags: [Support]
      summary: Add a message to a ticket
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "201": { description: Created } }

  # Service tiles beyond core Mobility
...
  /fix/requests/{id}/accept:
    post:
      tags: [Fix]
      summary: Artisan accepts a job
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /fix/requests/{id}/status:
    patch:
      tags: [Fix]
      summary: Update service request status
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /property/listings/{id}:
    get:
      tags: [Property]
      summary: Get one listing
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /property/listings/{id}/status:
    patch:
      tags: [Property]
      summary: Update listing status (owner)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /property/listings/{id}/enquiries:
    post:
      tags: [Property]
      summary: Enquire about a listing
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "201": { description: Created } }
    get:
      tags: [Property]
      summary: List enquiries on a listing (owner)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /admin/users/{id}:
    get:
      tags: [Admin]
      summary: Get one user
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /admin/drivers/{id}/verify:
    patch:
      tags: [Admin]
      summary: Approve/reject a driver application
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /admin/pricing/{vehicleType}:
    patch:
      tags: [Admin]
      summary: Update pricing rule for a vehicle type
      parameters: [{ name: vehicleType, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /admin/support/tickets/{id}/status:
    patch:
      tags: [Admin]
      summary: Update a support ticket's status
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /admin/artisans/{id}/verify:
    patch:
      tags: [Admin]
      summary: Approve/reject an artisan application
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
...
  /admin/property/listings/{id}/status:
    patch:
      tags: [Admin]
      summary: Moderate a property listing's status
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }